refactor(agents): split compaction planning projection
